[proc] Queue size based on message sets rather than per-message.
If we have hundreds of processes running on a box then we end up
flushing tens of messages per flush loop. This quickly fills up the
queue and gets us into a state where we can't actually read anything
without being less than the queue size.

As a fix we will push on the set of messages versus each message to the
`send` channel to match the designed behavior for this queue size
limiting. So now we only limit if we're backed up on a per-check basis
as expected.