kubernetes: Collect local pod and container state from kubelet
In an earlier change we cached the Kubernetes metadata state to avoid
additional load on a server. But this had the trade-off that we would
have out-of-date information which is definitely not great for
high-churn containers.

So as a solution this change will make is so we _always_ query the local
kubelet (which should be a very light, local query) for pod and
container state. This local data is merged with the global data to
provide both an up-to-date view as well as data that only comes from the
master (deployments, services).

The final trade-off is that we could have pods/containers that aren't
initially attached to their deployment or service until the next global
fetch.