Rework cgroup collection to determine pids from cgroup procs file.
This should limit the number of syscalls required by a good margin
because we will only read files for processes that are actually
controlled by a container. In the current version we check the cgroup
info for every live process on the system which leads to unnecesary
work and overhead.