Move containers to top-level of process messages, improve chunking. (#36)
* Move the containers to the top-level of both the CollectorProc and
CollectoRealTime messages. This is meant to match the new container
messsages and allows us to simplify processing on the backend.
* Also simplify and reorganize the message chunking logic. The goal is
to improve clarity and allow for re-usability of the container chunking
functions by the process checks.

Note that we're chunking containers up even if there are none in the
result but this should add little to no overhead.