Improve allocations in AllContainers
This is a call we will make many times over in general but especially in
the containers-only state which will be on by default.

The main source of allocations is because we have to read through every
`/proc/$pid/cgroup` file to find all valid cgroups and processes in
cgroups. So we've switched to smarter file reading where possible, but
there's certainly more we can do.

```
benchmark                    old ns/op     new ns/op     delta
BenchmarkAllContainers-2     4552072       3864796       -15.10%

benchmark                    old allocs     new allocs     delta
BenchmarkAllContainers-2     7571           2477           -67.28%
```