Add filtering for container name, image with k8s defaults.
Generally matching the behavior of the Agent 5 dockerutil, we will
filter out containers on the image or name. Unforunately, unlike the
Agent 5, we don't have a full set of tags for every container so
filtering beyond this would require more surgery.

By default we will filter out the noisy containers from Kubernetes like
the "pause" container.

As part of this we factored out the DockerConfig into a separate struct
since the arguments are likely going to grow here and it's nice to be
explicit about settings. And this allows us to decouple from the
AgentConfig so it can be used from other places.