[container] protect against nil stats on containers from ecs fargate (#176)
* [container] protect against nil stats on containers from ecs fargate

* Simplify nil safety on containers, add to all checks.

Switch back to storing the previous container from just the rate
metrics. This will simplfiy the interface for setting defaults on nil
containers and make it easy to apply it broadly.

The trade-off is maybe more memory but we've had this in all versions up
to this point so I'm not too worried about it.

Includes some simple tests that would panic before.