Use global network namespace if detected
When a netlink socket is created, messages for the current network
namespace are sent on the socket. This means that in a containernized
environment, we will by default get data for the containers network
namespace.

An easy workaround is to use `hostNetwork: true` when creating the
daemon. This isn't ideal.

This PR attempts to "find" the global namespace (which is reperesented
as a file descriptor) using the mounted host procfs and uses
go-conntrack's ability to temporarily change net namespace.

Requires the CAP_SYS_PTRACE capability, along with CAP_NET_ADMIN for
netlink.