STAC-13662 Don't limit inFlight max connections by the maxPerMessage. and bring back batching for connections (the receiving side can handle this now).
The InFlight limit can be a lot higher (default 10000), because connections will be filtered before sending them out. Moreover batching is reintroduced f
or connections so no need to have a high maxConnectionPerMessage limit anymore.

The maxPerMessage configuration is affected by the proc_limit amongst others. It does not make sense to limit the # connections in the outgoing data on a proc_limit.

Restoring the default maxPerMessage to 100, it was bumped to 2000 because chunking for processes and batching for connections was disabled (both are enabled again).